Restructured code by creating a single class Scale and subclasses Up/Down scale.
Approach: reduce to half as many times as possible, then crop such that
after cropping image size becomes a factor of the output image.

The factor has to be one of the following:
[3/4, 1/5, 2/5, 3/5, 4/5, 5/6, 4/7]

For a factor n/d, image is upsacled with an int factor n then downscaled
by an int factor d.

Firstly the required height is acheived followed by the same steps repeated
to acheive the required width.

The code is specifically written for an image of size (2592x1944) to be
downscaled to one of sizes: 2560x1440, 1920x1080, 1280x720, 640x480.